move stuff to qt
